Move tutorial under graphics and add appropriate tags.
