+++ Tutorial1-1Basics/src/main/res/values/strings.xml
@@ -47,7 +47,7 @@
     <string name="title_3_7_1">3-7-1 movableContentOf1</string>
     <string name="title_3_7_2">3-7-2 movableContentOf2</string>
     <string name="title_3_7_3">3-7-3 movableContentOf3</string>
-    <string name="title_3_7_4">3-7-4 CustomLayouCarousel</string>
+
     <string name="title_4_1"><![CDATA[4-1 Remember&MutableState]]></string>
     <string name="title_4_2_1">4-2-1 Recomposition</string>
     <string name="title_4_2_2">4-2-2 Recomposition2</string>
